# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIGR Angel Apparel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  description?: string;
  /** Price in cents */
  price: number;
  imageUrl?: string;
  inStock: boolean;
  /** Optional stock count. If null, stock is unlimited. Decrements on checkout. */
  stockCount?: number | null;
  category?: string;
  createdAt?: string;
}

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  inStock?: boolean;
  stockCount?: number | null;
  category?: string;
}

export type CheckoutRequestItemsItem = {
  productId: string;
  quantity: number;
};

export interface CheckoutRequest {
  items: CheckoutRequestItemsItem[];
  successUrl?: string;
  cancelUrl?: string;
}

export interface CheckoutResponse {
  url: string;
}

export interface NewsletterRequest {
  email: string;
}

export interface AdminLoginRequest {
  password: string;
}

export interface AdminLoginResponse {
  token: string;
}

export interface AdminSettings {
  stripePublishableKey?: string;
  stripeSecretKey?: string;
  hasStripeConfigured: boolean;
}

export interface UpdateAdminSettingsRequest {
  stripePublishableKey?: string;
  stripeSecretKey?: string;
  newPassword?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListProducts200 = {
  data: Product[];
};

export type GetStripePublicKey200 = {
  publishableKey: string;
};