# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIGR Angel Apparel API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all active products
 */
export const ListProductsResponse = zod.object({
  data: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      description: zod.string().optional(),
      price: zod.number().describe("Price in cents"),
      imageUrl: zod.string().optional(),
      inStock: zod.boolean(),
      stockCount: zod
        .number()
        .nullish()
        .describe(
          "Optional stock count. If null, stock is unlimited. Decrements on checkout.",
        ),
      category: zod.string().optional(),
      createdAt: zod.string().optional(),
    }),
  ),
});

/**
 * @summary Get a single product
 */
export const GetProductParams = zod.object({
  id: zod.coerce.string(),
});

export const GetProductResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number().describe("Price in cents"),
  imageUrl: zod.string().optional(),
  inStock: zod.boolean(),
  stockCount: zod
    .number()
    .nullish()
    .describe(
      "Optional stock count. If null, stock is unlimited. Decrements on checkout.",
    ),
  category: zod.string().optional(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Create a Stripe checkout session
 */
export const CreateCheckoutBody = zod.object({
  items: zod.array(
    zod.object({
      productId: zod.string(),
      quantity: zod.number(),
    }),
  ),
  successUrl: zod.string().optional(),
  cancelUrl: zod.string().optional(),
});

export const CreateCheckoutResponse = zod.object({
  url: zod.string(),
});

/**
 * @summary Subscribe to the newsletter
 */
export const SubscribeNewsletterBody = zod.object({
  email: zod.string(),
});

export const SubscribeNewsletterResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Developer admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Get admin settings (Stripe keys, password)
 */
export const GetAdminSettingsResponse = zod.object({
  stripePublishableKey: zod.string().optional(),
  stripeSecretKey: zod.string().optional(),
  hasStripeConfigured: zod.boolean(),
});

/**
 * @summary Update admin settings
 */
export const UpdateAdminSettingsBody = zod.object({
  stripePublishableKey: zod.string().optional(),
  stripeSecretKey: zod.string().optional(),
  newPassword: zod.string().optional(),
});

export const UpdateAdminSettingsResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Create a new product
 */
export const CreateAdminProductBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  imageUrl: zod.string().optional(),
  inStock: zod.boolean().optional(),
  stockCount: zod.number().nullish(),
  category: zod.string().optional(),
});

/**
 * @summary Update a product
 */
export const UpdateAdminProductParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateAdminProductBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  imageUrl: zod.string().optional(),
  inStock: zod.boolean().optional(),
  stockCount: zod.number().nullish(),
  category: zod.string().optional(),
});

export const UpdateAdminProductResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number().describe("Price in cents"),
  imageUrl: zod.string().optional(),
  inStock: zod.boolean(),
  stockCount: zod
    .number()
    .nullish()
    .describe(
      "Optional stock count. If null, stock is unlimited. Decrements on checkout.",
    ),
  category: zod.string().optional(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Delete a product
 */
export const DeleteAdminProductParams = zod.object({
  id: zod.coerce.string(),
});

export const DeleteAdminProductResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get public Stripe publishable key (for frontend checkout)
 */
export const GetStripePublicKeyResponse = zod.object({
  publishableKey: zod.string(),
});