# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIGR Angel Apparel API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  inStock?: boolean;
  stockCount?: number | null;
  category?: string;
}