# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIGR Angel Apparel API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  description?: string;
  /** Price in cents */
  price: number;
  imageUrl?: string;
  inStock: boolean;
  /** Optional stock count. If null, stock is unlimited. Decrements on checkout. */
  stockCount?: number | null;
  category?: string;
  createdAt?: string;
}